% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve metadata of a dataset from StatsWales}
\usage{
get_metadata(id)
}
\arguments{
\item{id}{A dataset id as a string}
}
\value{
If the dataset id is valid, the output will be
    the requested metadata in a dataframe. If the id is not
    valid, the function will return an HTTP error.
}
\description{
\code{get_metadata} returns a dataframe containing metadata from
\href{https://statswales.gov.wales}{StatsWales} using a dataset id.
}
\examples{
metadata <- get_metadata("hlth0515")

}
