% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statswales_get_metadata.R
\name{statswales_get_metadata}
\alias{statswales_get_metadata}
\title{Retrieve metadata of a dataset from StatsWales}
\usage{
statswales_get_metadata(id, language = "english")
}
\arguments{
\item{id}{A dataset id as a string}

\item{language}{A string. Returns the metadata in either English ('english') or Welsh ('welsh').
The default is English.}
}
\value{
If the dataset id is valid, the output will be
the requested metadata in a dataframe. If the id is not
valid, the function will return an HTTP error.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because StatsWales ended its OData
service in August 2024.
}
\details{
\code{statswales_get_metadata} returns a dataframe containing metadata from
\href{https://statswales.gov.wales}{StatsWales} using a dataset id.
}
\examples{
metadata <- statswales_get_metadata("hlth0515")

}
\keyword{internal}
