% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statswales_search.R
\name{statswales_search}
\alias{statswales_search}
\title{Search StatsWales datasets}
\usage{
statswales_search(search_text, language = "english")
}
\arguments{
\item{search_text}{A vector of search terms.}

\item{language}{A string. Returns the metadata in either English ('english') or Welsh ('welsh')}
}
\value{
A dataframe of StatsWales dataset titles and IDs.
}
\description{
\code{statswales_search} returns a dataframe of dataset titles and IDs from
\href{https://statswales.gov.wales}{StatsWales}, based on the user's
text input.
}
\examples{
crops_datasets <- statswales_search("*crops*")



}
