% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statswalesr-package.R
\docType{package}
\name{statswalesr-package}
\alias{statswalesr}
\alias{statswalesr-package}
\title{statswalesr: Easily Extract Data from 'StatsWales'}
\description{
Download data from the 'StatsWales' public API (\url{https://api.stats.gov.wales/v2}) into R. Provides functions to list, search, and browse datasets by topic, retrieve data with optional filtering, sorting, and pivoting, and download datasets as CSV or Excel files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jamie-ralph/statswalesr}
  \item \url{https://jamie-ralph.github.io/statswalesr/}
  \item Report bugs at \url{https://github.com/jamie-ralph/statswalesr/issues}
}

}
\author{
\strong{Maintainer}: Jamie Ralph \email{jamesryanralph@outlook.com}

Authors:
\itemize{
  \item Jamie Ralph \email{jamesryanralph@outlook.com}
}

}
\keyword{internal}
