% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statswalesr-package.R
\docType{package}
\name{statswalesr-package}
\alias{statswalesr}
\alias{statswalesr-package}
\title{statswalesr: Easily Extract Data from 'StatsWales'}
\description{
Download data from 'StatsWales' into R.
}
\author{
\strong{Maintainer}: Jamie Ralph \email{jamesryanralph@outlook.com}

}
\keyword{internal}
