% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statswalesr-package.R
\docType{package}
\name{statswalesr-package}
\alias{statswalesr}
\alias{statswalesr-package}
\title{statswalesr: Easily Extract Data from 'StatsWales'}
\description{
Download data from 'StatsWales' into R. Removes the need for the user to write their own loops when parsing data from the 'StatsWales' API. Provides functions for datasets (\url{http://open.statswales.gov.wales/en-gb/dataset}) and metadata (\url{http://open.statswales.gov.wales/en-gb/discover/metadata}) endpoints.
}
\author{
\strong{Maintainer}: Jamie Ralph \email{jamesryanralph@outlook.com}

}
\keyword{internal}
